chore: move the quote stream complete banner to above the token warning banner
